% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire damage and proximity to a fire station}
\format{
A dataframe with 15 rows and 2 variables:
\describe{
\item{DISTANCE}{Closest distance in miles to a fire station}
\item{DAMAGE}{Damage in thousands of dollars caused by fire}
}
}
\source{
Statistics for Engineering and the Sciences, 6th Edition
}
\usage{
fire
}
\description{
A dataset with two quantitative variables
Interest is in the relationship between proximity to a fire station and damage
}
\keyword{datasets}
