% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{The number of iterations of samples}

\item{n}{The size of the sample}

\item{p}{The probability of a success}
}
\value{
A barplot of the proportions of the number of successes
}
\description{
mybin
}
\examples{
\dontrun{mybin(100,20,0.7)}
}
