% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{The number of resamples to be taken}

\item{x}{The sample from which resampling will be done}

\item{fun}{The function that will calculate the desired statistic from the samples}

\item{alpha}{(1-alpha)100% defines the confidence interval to be calculate}

\item{cx}{Plot scaling for the text}

\item{...}{Any other parameters to be given to the plot}
}
\value{
A histogram of the given statistic with confidence interval annotations and a list with the confidence intervals, function, and sample data
}
\description{
myboot2
}
\examples{
\dontrun{myboot2(iter=10000,x=sam,fun="var",alpha=0.2)}
}
