% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{mycltb}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{The size of the binomial samples}

\item{iter}{The number of binomial samples}

\item{p}{The probability of a success}

\item{...}{Any graphical options to add to the histogram}
}
\value{
A histogram of the relative frequencies of the sample means
}
\description{
mycltb
}
\examples{
\dontrun{mycltb(n=5,iter=10000,p=0.3)}
}
