% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{myddt}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{The dataframe to be manipulated (Only works for DDT)}

\item{SPECIES}{The species of fish from DDT to be plotted}
}
\value{
A plot of length vs weight of the given species and a list of the original data, the modified data, and the relative frequency of rivers
}
\description{
myddt
}
\examples{
\dontrun{myddt(df = ddt, SPECIES = "CCATFISH")}
}
