% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{The logarithmic function used to estimate the maximum likelihood of the parameter}

\item{x}{The sample of data}

\item{param}{An interval of values to be used as the desired parameter}

\item{...}{Any additional graphical parameters desired}
}
\value{
A plot of the log likelihood with the maximum labeled, the value of the maximum likelihood parameter estimate, the maximum likelihood, and a vector of slopes before and after the parameter estimate
}
\description{
mymaxlik
}
\examples{
\dontrun{mymaxlik(x=y,lfun="mean",param=seq(0,1,length=1000))}
}
