% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the normal distribution}

\item{sigma}{The standard deviation of the normal distribution}

\item{a}{The right bound of the probability to shaded and calculate (Left bound is -infinity)}
}
\value{
A graph of the normal distribution with the specified probability shaded
}
\description{
myncurve
}
\examples{
\dontrun{myncurve(5,2,4)}
}
