% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{Read in data}
\usage{
myread(csv)
}
\arguments{
\item{csv}{The file to be read in}
}
\value{
A data frame object
}
\description{
Read in data
}
\examples{
\dontrun{mpg.df=myread("EPAGAS.csv")}
}
